requirements.txt
azure-servicebus==7.8.0
python-dotenv==0.20.0
requests==2.27.1
urllib3==1.26.5
